Created Disconnection.java class to pull up method and variables from the DisconnectingEvent and DisconnectedEvent classes
